@@ -1,8 +1,8 @@
-const { evaluateCEL } = require('./index.js');
+import { evaluateCEL } from '../dist/index.js';
 
 async function main() {
   console.log('CEL Expression Evaluator Example\n');
-  console.log('=' .repeat(50));
+  console.log('='.repeat(50));
 
   // Example 1: Basic arithmetic
   console.log('\n1. Basic Arithmetic:');
@@ -28,7 +28,10 @@ async function main() {
   // Example 3: Comparison
   console.log('\n3. Comparison:');
   try {
-    const result3 = await evaluateCEL('x > y ? "greater" : "lesser"', { x: 10, y: 5 });
+    const result3 = await evaluateCEL('x > y ? "greater" : "lesser"', {
+      x: 10,
+      y: 5,
+    });
     console.log('   Expression: x > y ? "greater" : "lesser"');
     console.log('   Variables: { x: 10, y: 5 }');
     console.log('   Result:', result3.result);
@@ -39,11 +42,16 @@ async function main() {
   // Example 4: String operations
   console.log('\n4. String Operations:');
   try {
-    const result4 = await evaluateCEL('name + " is " + string(age) + " years old"', {
-      name: 'Alice',
-      age: 30
-    });
-    console.log('   Expression: name + " is " + string(age) + " years old"');
+    const result4 = await evaluateCEL(
+      'name + " is " + string(age) + " years old"',
+      {
+        name: 'Alice',
+        age: 30,
+      },
+    );
+    console.log(
+      '   Expression: name + " is " + string(age) + " years old"',
+    );
     console.log('   Variables: { name: "Alice", age: 30 }');
     console.log('   Result:', result4.result);
   } catch (err) {
@@ -53,7 +61,9 @@ async function main() {
   // Example 5: List operations
   console.log('\n5. List Operations:');
   try {
-    const result5 = await evaluateCEL('myList.size() > 0', { myList: [1, 2, 3, 4, 5] });
+    const result5 = await evaluateCEL('myList.size() > 0', {
+      myList: [1, 2, 3, 4, 5],
+    });
     console.log('   Expression: myList.size() > 0');
     console.log('   Variables: { myList: [1, 2, 3, 4, 5] }');
     console.log('   Result:', result5.result);
@@ -64,10 +74,15 @@ async function main() {
   // Example 6: Map operations
   console.log('\n6. Map Operations:');
   try {
-    const result6 = await evaluateCEL('user["name"] + " has " + string(user["score"]) + " points"', {
-      user: { name: 'Bob', score: 100 }
-    });
-    console.log('   Expression: user["name"] + " has " + string(user["score"]) + " points"');
+    const result6 = await evaluateCEL(
+      'user["name"] + " has " + string(user["score"]) + " points"',
+      {
+        user: { name: 'Bob', score: 100 },
+      },
+    );
+    console.log(
+      '   Expression: user["name"] + " has " + string(user["score"]) + " points"',
+    );
     console.log('   Variables: { user: { name: "Bob", score: 100 } }');
     console.log('   Result:', result6.result);
   } catch (err) {
@@ -80,7 +95,7 @@ async function main() {
     const result7 = await evaluateCEL('(x > 0 && y > 0) || z > 100', {
       x: 5,
       y: 10,
-      z: 50
+      z: 50,
     });
     console.log('   Expression: (x > 0 && y > 0) || z > 100');
     console.log('   Variables: { x: 5, y: 10, z: 50 }');
@@ -93,7 +108,7 @@ async function main() {
   console.log('Examples completed!');
 }
 
-main().catch(err => {
+main().catch((err) => {
   console.error('Fatal error:', err);
   process.exit(1);
 });
